pyglass/requirements.txt
black
dill
findiff
ipykernel
matplotlib
mayavi
numba
numpy
pip-search
plyfile
pyflakes
pyyaml
PyQt5
python-lsp-server[all]
scipy
sympy
scikit-image
